Improve UI and move folder
- Change colours in the app to follow the chosen colour palette
- Increase the font size of most text to fulfil a user story
- Change minor parts of UI
- Move controller to UI as it is part of UI
