+++ src/test/java/seedu/address/logic/commands/DeleteCommandTest.java
@@ -14,12 +14,13 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.controller.ConfirmationBypassController;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
+import seedu.address.ui.controller.ConfirmationBypassController;
+
 
 /**
  * Contains integration tests (interaction with the Model) and unit tests for
